requirements.txt
numpy==1.26.4
opencv-python==4.10.0.82
pandas==2.2.2
requests==2.32.3
pyserial==3.5